Move code search api call to its own file
