+++ extensions/ql-vscode/src/databases/ui/db-panel.ts
@@ -17,7 +17,6 @@ import {
 import {
   showAndLogErrorMessage,
   showAndLogInformationMessage,
-  showAndLogWarningMessage,
 } from "../../helpers";
 import { DisposableObject } from "../../pure/disposable-object";
 import {
@@ -37,11 +36,8 @@ import { getControllerRepo } from "../../variant-analysis/run-remote-query";
 import { getErrorMessage } from "../../pure/helpers-pure";
 import { DatabasePanelCommands } from "../../common/commands";
 import { App } from "../../common/app";
-import { getCodeSearchRepositories } from "../../variant-analysis/gh-api/gh-api-client";
 import { QueryLanguage } from "../../common/query-language";
-import { retry } from "@octokit/plugin-retry";
-import { throttling } from "@octokit/plugin-throttling";
-import { Octokit } from "@octokit/rest";
+import { getCodeSearchRepositories } from "../code-search-api";
 
 export interface RemoteDatabaseQuickPickItem extends QuickPickItem {
   remoteDatabaseKind: string;
@@ -409,7 +405,7 @@ export class DbPanel extends DisposableObject {
           `${codeSearchQuery} ${languagePrompt}`,
           progress,
           token,
-          await this.getOctokitForSearch(),
+          this.app.credentials,
         );
 
         token.onCancellationRequested(() => {
@@ -502,31 +498,5 @@ export class DbPanel extends DisposableObject {
         )}`,
       );
     }
-  }
-
-  // since we don't have access to the extensions log methods we initialize octokit here
-  private async getOctokitForSearch(): Promise<Octokit> {
-    const MyOctokit = Octokit.plugin(throttling);
-    const auth = await this.app.credentials.getAccessToken();
-
-    const octokit = new MyOctokit({
-      auth,
-      retry,
-      throttle: {
-        onRateLimit: (retryAfter: number, options: any): boolean => {
-          void showAndLogWarningMessage(
-            `Request quota exhausted for request ${options.method} ${options.url}. Retrying after ${retryAfter} seconds!`,
-          );
-
-          return true;
-        },
-        onSecondaryRateLimit: (_retryAfter: number, options: any): void => {
-          void showAndLogWarningMessage(
-            `SecondaryRateLimit detected for request ${options.method} ${options.url}`,
-          );
-        },
-      },
-    });
-    return octokit;
   }
 }
