+++ extensions/ql-vscode/src/variant-analysis/gh-api/gh-api-client.ts
@@ -7,44 +7,6 @@ import {
   VariantAnalysisSubmissionRequest,
 } from "./variant-analysis";
 import { Repository } from "./repository";
-import { Progress } from "vscode";
-import { CancellationToken } from "vscode-jsonrpc";
-import { Octokit } from "@octokit/rest";
-
-export async function getCodeSearchRepositories(
-  query: string,
-  progress: Progress<{
-    message?: string | undefined;
-    increment?: number | undefined;
-  }>,
-  token: CancellationToken,
-  octokit: Octokit,
-): Promise<string[]> {
-  let nwos: string[] = [];
-
-  for await (const response of octokit.paginate.iterator(
-    octokit.rest.search.code,
-    {
-      q: query,
-      per_page: 100,
-    },
-  )) {
-    nwos.push(...response.data.map((item) => item.repository.full_name));
-    // calculate progress bar: 80% of the progress bar is used for the code search
-    const totalNumberOfRequests = Math.ceil(response.data.total_count / 100);
-    // Since we have a maximum of 1000 responses of the api, we can use a fixed increment whenever the totalNumberOfRequests would be greater than 10
-    const increment =
-      totalNumberOfRequests < 10 ? 80 / totalNumberOfRequests : 8;
-    progress.report({ increment });
-
-    if (token.isCancellationRequested) {
-      nwos = [];
-      break;
-    }
-  }
-
-  return [...new Set(nwos)];
-}
 
 export async function submitVariantAnalysis(
   credentials: Credentials,
